move Day to arrow-ui
