+++ modules/docs/arrow-examples/build.gradle
@@ -27,6 +27,7 @@ dependencies {
     compile project(':arrow-effects-extensions')
     compile project(':arrow-effects-rx2-extensions')
     compile project(':arrow-optics')
+    compile project(':arrow-ui')
     compile project(':arrow-validation')
 
     kapt project(':arrow-meta')
