+++ modules/ui/arrow-ui/gradle.properties
@@ -1,0 +1,4 @@
+# Maven publishing configuration
+POM_NAME=Arrow UI
+POM_ARTIFACT_ID=arrow-ui
+POM_PACKAGING=jar
