+++ modules/ui/arrow-ui/src/test/kotlin/arrow/ui/DayTest.kt
@@ -1,11 +1,11 @@
-package arrow.data
+package arrow.ui
 
 import arrow.core.ForId
 import arrow.core.Id
 import arrow.core.Tuple2
 import arrow.core.Tuple2Of
-import arrow.data.extensions.day.applicative.applicative
-import arrow.data.extensions.day.comonad.comonad
+import arrow.ui.extensions.day.applicative.applicative
+import arrow.ui.extensions.day.comonad.comonad
 import arrow.core.extensions.id.applicative.applicative
 import arrow.core.extensions.id.comonad.comonad
 import arrow.core.fix
